requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
docopt==0.6.2
idna==2.8
lxml==4.4.2
num2words==0.5.10
pdf2image==1.11.0
Pillow==6.2.1
psutil==5.6.7
pycryptodome==3.9.4
PyPDF2==1.26.0
pytesseract==0.3.1
requests==2.22.0
selenium==3.141.0
soupsieve==1.9.5
urllib3==1.25.7
